+++ pages/lists/[listId]/contributors.tsx
@@ -1,9 +1,6 @@
-import useSWR, { Fetcher } from "swr";
-import { useState } from "react";
 import { GetServerSidePropsContext } from "next";
 import { createPagesServerClient } from "@supabase/auth-helpers-nextjs";
 import ListPageLayout from "layouts/lists";
-import publicApiFetcher from "lib/utils/public-api-fetcher";
 import ContributorTable from "components/organisms/ContributorsTable/contributors-table";
 import PaginationResults from "components/molecules/PaginationResults/pagination-result";
 import ContributorListTableHeaders from "components/molecules/ContributorListTableHeader/contributor-list-table-header";
@@ -11,60 +8,7 @@ import Pagination from "components/molecules/Pagination/pagination";
 import { fetchApiData, validateListPath } from "helpers/fetchApiData";
 import Error from "components/atoms/Error/Error";
 import useSupabaseAuth from "lib/hooks/useSupabaseAuth";
-
-interface ContributorList {
-  data: DBListContributor[];
-  meta: Meta;
-}
-
-function convertToContributors(rawContributors: ContributorList["data"] = []): DbPRContributor[] {
-  const contributors = rawContributors
-    ? rawContributors.map((contributor) => {
-        return {
-          author_login: contributor.login,
-          updated_at: contributor.created_at,
-          user_id: contributor.id,
-        };
-      })
-    : [];
-
-  return contributors;
-}
-
-const useContributorsList = ({
-  listId,
-  initialData,
-  initialPage = 1,
-  defaultLimit = 10,
-}: {
-  listId: string;
-  initialData?: {
-    data: DbPRContributor[];
-    meta: Meta;
-  };
-  initialPage?: number;
-  defaultLimit?: number;
-}) => {
-  const [page, setPage] = useState(initialPage);
-  const [limit, setLimit] = useState(defaultLimit);
-  const { data, error, mutate } = useSWR<any>(
-    `lists/${listId}/contributors?page=${page}&limit=${limit}`,
-    publicApiFetcher as Fetcher<ContributorList, Error>,
-    {
-      fallbackData: initialData,
-    }
-  );
-  const contributors = convertToContributors(data?.data);
-
-  return {
-    setPage,
-    setLimit,
-    data: data ? { data: contributors, meta: data.meta } : { data: [], meta: {} },
-    isLoading: !error && !data,
-    isError: !!error,
-    mutate,
-  };
-};
+import { convertToContributors, useContributorsList } from "lib/hooks/api/useContributorList";
 
 export const getServerSideProps = async (ctx: GetServerSidePropsContext) => {
   const supabase = createPagesServerClient(ctx);
@@ -77,7 +21,7 @@ export const getServerSideProps = async (ctx: GetServerSidePropsContext) => {
   const { listId } = ctx.params as { listId: string };
   const limit = 10; // Can pull this from the querystring in the future
   const [{ data, error: contributorListError }, { data: list, error }] = await Promise.all([
-    fetchApiData<ContributorList>({
+    fetchApiData<PagedData<DBListContributor>>({
       path: `lists/${listId}/contributors?limit=${limit}`,
       bearerToken,
       pathValidator: validateListPath,
